Move files to util folder
