+++ augur/api/routes/user.py
@@ -13,7 +13,7 @@ from sqlalchemy.sql import text
 from sqlalchemy.orm import sessionmaker
 from augur.application.db.session import DatabaseSession
 from augur.tasks.github.util.github_task_session import GithubTaskSession
-from augur.application.cli._repo_load_controller import RepoLoadController
+from augur.util.repo_load_controllerimport RepoLoadController
 
 
 from augur.application.db.models import User
