+++ tests/test_applicaton/test_repo_load_controller/test_repo_load_controller.py
@@ -4,7 +4,7 @@ import uuid
 import sqlalchemy as s
 
 
-from augur.application.cli._repo_load_controller import RepoLoadController
+from augur.util.repo_load_controllerimport RepoLoadController
 
 from augur.tasks.github.util.github_task_session import GithubTaskSession
 from augur.application.db.session import DatabaseSession
